libavformat/rtmp: Implements RTMP reconnect feature
Nowadays when you are streaming to a live platform if the RTMP(s)
server needs to restarted for any reason (ex: deploy new version)
the RTMP connection is interrupted (probably after some draining time).
Facebook will publish a proposal to avoid that by sending a
GoAway message in the RTMP protocol.
This code is the reference client implementation of that proposal.
AFAIK other big live platforms showed their interest in implementing
this mechanism.
This can be already tested against Facebook live production using
the querystring parameter ?ccr_sec=120 (that indicates the backend
to send a disconnect signal after those seconds)